# tracer/restapi/api.yaml
swagger: '2.0'
info:
  title: Goapptrace REST API
  description: '**WARNING**: This API specification is draft version.'
  version: '0.1'
  license:
    name: MIT
    url: 'https://opensource.org/licenses/MIT'
basePath: /api/v0.1
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
definitions:
...
  server:
    description: Status of a log server.
    type: object
    required:
      - server-id
      - version
      - address
    properties:
      server-id:
        type: integer
        format: int64
        minimum: 1
        example: 10
      version:
        description: >-
          Version of the status. API server should increase version field value
          when this status was updated.
        type: integer
        format: int64
        minimum: 1
        example: 128
      address:
        description: Host and port.
        type: string
        pattern: '^[a-zA-Z0-9\-]+:\d+$'
        example: 'localhost:9000'
...
  log:
    description: Status of log file.
    type: object
    required:
      - log-id
      - version
      - max-file-size
      - read-only
    properties:
      log-id:
        type: string
        example: f459a84959e23d643705c8d6df19f4d0
        minLength: 32
        maxLength: 32
        description: LogID is string like hex value.
      version:
        type: integer
        format: int64
        minimum: 1
        example: 13
      max-file-size:
        type: integer
        format: int64
        description: separate the log file when file size exceeds max-file-size.
      read-only:
        type: boolean
      metadata:
        $ref: '#/definitions/log-metadata'
...
  func-call-jsonlines:
    description: The multiple json separated by newline character.
    type: array
    items:
      $ref: '#/definitions/func-call'
...
  goroutine-jsonlines:
    description: The multiple json separated by newline character.
    type: array
    items:
      $ref: '#/definitions/goroutine'
...
paths:
...
  '/server/{server-id}/status':
    parameters:
      - name: server-id
        in: path
        required: true
        type: integer
    get:
      description: Returns current status of a server.
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/server'
    put:
      description: Updates current status of a server.
      parameters:
        - name: version
          in: query
          description: Old status version.
          required: true
          type: integer
        - name: status
          in: body
          description: New status.
          required: true
          schema:
            $ref: '#/definitions/server'
      responses:
        '200':
          description: Successed to update status.
          schema:
            $ref: '#/definitions/server'
        '409':
          description: >-
            Failed to update status because of conflict. The API Server returns
            latest status. The API client had better retry to update status.
          schema:
            $ref: '#/definitions/server'
  '/server/{server-id}/watch':
    get:
      description: >-
        Long-polling API for status update notification. This API will be
        waiting for the status to change or timeout.
      parameters:
        - name: server-id
          in: path
          required: true
          type: integer
        - name: version
          in: query
          description: Status version.
          required: true
          type: integer
        - name: timeout
          in: query
          description: The unit is seconds.
          required: true
          type: integer
      responses:
        '200':
          description: Notify of status change.
          schema:
            $ref: '#/definitions/server'
        '204':
          description: Notify that this status was not changed until timeout.
...
  '/log/{log-id}':
    delete:
      description: Delete a log file.
      parameters:
        - name: log-id
          in: path
          required: true
          type: integer
      responses:
        '204':
          description: OK
    get:
      description: Returns current status of a specified log.
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/log'
    put:
      description: Updates current status of a server.
      parameters:
        - name: version
          in: query
          description: Old status version.
          required: true
          type: integer
        - name: status
          in: body
          description: New status.
          schema:
            $ref: '#/definitions/log'
      responses:
        '200':
          description: Successed to update status.
          schema:
            $ref: '#/definitions/log'
        '409':
          description: >-
            Failed to update status because of conflict. The API Server returns
            latest status. The API client had better retry to update status.
          schema:
            $ref: '#/definitions/log'
  '/log/{log-id}/watch':
    get:
      description: >-
        Long-polling API for status update notification. This API will be
        waiting for the status to change or timeout.
      parameters:
        - name: log-id
          in: path
          required: true
          type: integer
        - name: version
          in: query
          description: Status version.
          required: true
          type: integer
        - name: timeout
          in: query
          description: The unit is seconds.
          required: true
          type: integer
      responses:
        '200':
          description: Notify for status change.
          schema:
            $ref: '#/definitions/log'
        '204':
          description: Notify for non-changed in timeout.
  '/log/{log-id}/func-call/search':
    get:
      description: Filters the function call log records.
      produces:
        - application/x-jsonlines
      parameters:
        - name: log-id
          in: path
          required: true
          type: integer
        - name: txid
          in: query
          description: Transaction ID.
          type: string
        - name: gid
          in: query
          description: Goroutine ID.
          type: integer
        - name: fid
          in: query
          description: Function ID.
          type: integer
        - name: mid
          in: query
          description: Module ID.
          type: integer
        - name: min-id
          in: query
          description: Minimum of log ID.
          type: integer
        - name: max-id
          in: query
          description: Maximum of log ID.
          type: integer
        - name: min-timestamp
          in: query
          description: Minimum of timestamp.
          type: integer
        - name: max-timestamp
          in: query
          description: Maximum of timestamp.
          type: integer
        - name: limit
          in: query
          description: Maximum number of results. I recommend that specify parameters of "sort" and "order" in addition to "limit" parameter.
          type: integer
        - name: sort
          in: query
          description: Sort key. This parameter takes only "id" or "start-time" or "end-time".
          type: string
        - name: order
          in: query
          description: Sort order. This parameter takes only of “asc” (default) and “desc”.
          type: string
      responses:
        '200':
          description: Returns function call log records.
          schema:
            $ref: '#/definitions/func-call-jsonlines'
  '/log/{log-id}/func-call/stream':
    get:
      description: >-
        ilters the function call log records. This API and search API are
        basically the same. The difference point is behavior to new records
        received after API call. Search API returns a response immediately, and
        the response are not include new records. Stream API sending a response
        little by little including new records.
      produces:
        - application/x-jsonlines
      parameters:
        - name: log-id
          in: path
          required: true
          type: integer
        - name: txid
          in: query
          description: Transaction ID.
          type: string
        - name: gid
          in: query
          description: Goroutine ID.
          type: integer
        - name: fid
          in: query
          description: Function ID.
          type: integer
        - name: mid
          in: query
          description: Module ID.
          type: integer
        - name: min-id
          in: query
          description: Minimum of log ID.
          type: integer
        - name: max-id
          in: query
          description: Maximum of log ID.
          type: integer
        - name: min-timestamp
          in: query
          description: Minimum of timestamp.
          type: integer
        - name: max-timestamp
          in: query
          description: Maximum of timestamp.
          type: integer
      responses:
        '200':
          description: Returns log records forever.
          schema:
            $ref: '#/definitions/func-call-jsonlines'
  '/log/{log-id}/goroutines/search':
    get:
      description: Returns list of goroutines.
      produces:
        - application/x-jsonlines
      parameters:
        - name: log-id
          in: path
          required: true
          type: integer
        - name: min-timestamp
          in: query
          description: Minimum of timestamp.
          type: integer
        - name: max-timestamp
          in: query
          description: Maximum of timestamp.
          type: integer
      responses:
        '200':
          description: success
          schema:
            $ref: '#/definitions/goroutine-jsonlines'
  '/log/{log-id}/symbols':
    get:
      description: Returns symbols.
      parameters:
        - name: log-id
          in: path
          required: true
          type: integer
      responses:
        '200':
          description: success
          schema:
            $ref: '#/definitions/symbols'
  '/log/{log-id}/symbol/module/{pc}':
    get:
      description: Returns detail of the module.
      parameters:
        - name: log-id
          in: path
          required: true
          type: integer
        - name: pc
          in: path
          required: true
          type: integer
      responses:
        '200':
          description: success
          schema:
            $ref: '#/definitions/symbol-module'
  '/log/{log-id}/symbol/func/{pc}':
    get:
      description: Returns detail of the function.
      parameters:
        - name: log-id
          in: path
          required: true
          type: integer
        - name: pc
          in: path
          required: true
          type: integer
      responses:
        '200':
          description: success
          schema:
            $ref: '#/definitions/symbol-func'
  '/log/{log-id}/symbol/line/{pc}':
    get:
      description: Returns detail of the line.
      parameters:
        - name: log-id
          in: path
          required: true
          type: integer
        - name: pc
          in: path
          required: true
          type: integer
      responses:
        '200':
          description: success
          schema:
            $ref: '#/definitions/symbol-line'
...
  '/tracer/{tracer-id}':
    delete:
      description: Delete a tracer. This tracer will stop forcibly.
      parameters:
        - name: tracer-id
          in: path
          required: true
          type: integer
      responses:
        '202':
          description: >-
            Request is accepted. This response do not guarantee that this tracer
            is stopped.
...
  '/tracer/{tracer-id}/watch':
    get:
      description: >-
        Long-polling API for status update notification. This API will be
        waiting for the status to change or timeout.
      parameters:
        - name: tracer-id
          in: path
          required: true
          type: integer
        - name: version
          in: query
          description: Status version.
          required: true
          type: integer
        - name: timeout
          in: query
          description: The unit is seconds.
          required: true
          type: integer
      responses:
        '200':
          description: Notify for status change.
          schema:
            $ref: '#/definitions/server'
        '204':
          description: Notify for non-changed in timeout.